src/agent/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asgiref
asttokens
attrs
backoff
bcrypt
beautifulsoup4
bs4
cachetools
certifi
charset-normalizer
chroma-hnswlib
chromadb==0.4.9
click
coloredlogs
dataclasses-json
decorator
Deprecated
fastapi
filelock
flask
flask-cors
flatbuffers
frozenlist
fsspec
google-auth
grpcio
h11
httpcore
httptools
httpx
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
ipython
jedi
jsonpatch
jsonpointer
kubernetes
langchain
langchain-chroma
langchain-community
langchain-core
langchain-nvidia-ai-endpoints
langchain-text-splitters
langgraph-checkpoint
langsmith
langgraph
marshmallow
matplotlib-inline
mdurl
multidict
mypy-extensions
numpy
oauthlib
onnxruntime
orjson
pydantic
pydantic-settings
pydantic_core
Pygments
pypdf
python-dotenv
PyYAML
requests
rich
SQLAlchemy
starlette
streamlit
tenacity
tiktoken
tokenizers
tqdm
traitlets
typing-inspect
typing_extensions
urllib3
uvicorn
watchfiles
websocket-client
websockets
wrapt
yarl
zipp
arxiv
wikipedia
pymupdf
requests